movies_api/pyproject.toml
[project]
name = "movies-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "elasticsearch[async]>=9.3.0",
    "fastapi[all]>=0.135.1",
    "gunicorn>=25.3.0",
    "pyjwt>=2.12.1",
    "python-dateutil>=2.9.0.post0",
    "redis>=7.3.0",
]

[dependency-groups]
dev = [
    "faker>=40.12.0",
    "isort>=8.0.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.6",
    "ty>=0.0.23",
]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py314"
exclude = [
    ".venv"
]

[tool.ruff.lint]
preview = true
select = ["E", "F", "B", "COM", "N", "Q", "LOG", "G", "RET", "I", "ANN"]
ignore = ["E203", "I001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
line_length = 88
skip = [
    ".venv"
]


[tool.ty.src]
exclude = [
    ".venv"
]


[tool.pytest.ini_options]
asyncio_default_test_loop_scope = "session"

